+++ zarg.h
@@ -11,11 +11,13 @@ typedef struct {
     char *description;
 } Flag;
 
+bool valid_flag(char *argument);
+bool match_flag(char *argument, Flag *flag);
+
 int    flag_count(char **arguments, Flag *flag);
 char **flag_value(char **arguments, Flag *flag);
 
 int  ppclen(char **array);
-bool valid_flag(char *argument);
-bool match_flag(char *argument, Flag *flag);
+bool zinit(char **argv, Flag flags[], int length);
 
 #endif